Wow, you really do need those volatile keywords!
For unsynchronized sections that were resetting the various timeouts
from different threads this was resulting in noticable quirks such as
double timeouts, etc pretty reliably. Seemed like the variable was not
reflecting an appropriate state and thus couldn't be cancelled.

The thread safety on these classes really has become a mess and
definitely needs to be cleaned up.