Moved RaftNode operations to a single thread
This contains the functioning of an entire send/receive which helps
operations like node conversions be atomic before moving on. I'm not
sure that this will be very efficient when doing things like appending
entries to many nodes. It does seem to keep things consistent, though.

Maybe what it needs is to have a lock that is owned by the RaftNode to
survive node mode transitions?